feat: 여행지 상세조회 디자인

+++ src/components/plan/PlanListView.vue
@@ -30,6 +30,8 @@ const close = () => {
 };
 
 const removePlan = async (id) => {
+  const answer = confirm("정말 삭제하시겠습니까?");
+  if (!answer) return;
   await planApi.delete(`/${id}`);
   getPlanList();
 };
@@ -54,7 +56,7 @@ getPlanList();
         :key="plan.id"
         class="m-2 group px-10 py-5 bg-white/10 rounded-lg flex flex-col items-center justify-center gap-2 relative after:absolute after:h-full after:bg-[#abd373] z-20 shadow-lg after:-z-20 after:w-full after:inset-0 after:rounded-lg transition-all duration-300 hover:transition-all hover:duration-300 after:transition-all after:duration-500 after:hover:transition-all after:hover:duration-500 overflow-hidden cursor-pointer after:-translate-y-full after:hover:translate-y-0 [&amp;_p]:delay-200 [&amp;_p]:transition-all"
         @click="detailPlan(plan)"
-        >
+      >
         <div class="image-container">
           <img
             class="planImg"
@@ -194,7 +196,6 @@ getPlanList();
           </p>
           <p
             class="btun4 lg:inline-flex items-center gap-3 group-hover:bg-white/10 bg-[#abd373] shadow-[10px_10px_150px_#ff9f0d] cursor-pointer py-2 px-4 text-sm font-semibold rounded-full butn"
-            
           >
             더보기
           </p>
